Removed final dependency on icu
